requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# upload_to_drive.py: 4,5
google_api_python_client == 1.8.3

# upload_to_drive.py: 7
google_auth == 1.14.3

# upload_to_drive.py: 6
google_auth_oauthlib == 0.4.1
